Moved position of score and reset button
Move position of score and reset button on html page.  Comment out
code not used in script.js.  Move the position of const startButton
to the top with the rest of the const variables. Alter function
enterGame code by removing function at line 73 and renamed scoreValue
to score.  Revised display function at the end of the js file.