Fix styling elements on page
Center buttons for Start Game and Reset placed them in the one
div in index.html and added styling for them in style.css.
Also adjusted time remaining and score by putting both into the one
div in index.html and added styling to center them on the page and
changed color to red and blue. In index.html placed the
square-container in a main-container to center it on the page.
Also add css styling to change background color to dark grey to
contract with buttons and banner and added box shadow of yellow
just to give a nice effect to each circle.  Increased the widht
of the darkgray banner at the top of the page.