Revised all functions in script.js
In script.js Removed commented out code, put the majority of
variables togeather at the top of file. Removed code that was not
working code.  Refdefined functions in order to get cat displaying
in each circle randomly and displaying when the timer started to
count down. Fix the score so that it displays to the  user on the page.