Add functionality to show selected game level
Add functionality so that when a user selects a game
level that this is hightlighted so the user knows what
level they are playing also when the page is loaded the
game will automatically default to the levelEasy and the
timer will reflect same.  Added classes to button section
in html so they could be styled in style.css file.  in
script.js file add event listeners for level buttons and
function to initialize game and when level button is clicked
to hightlight that button so user knows what level they are
playing at.