fix position of cats in each circle
fix position of cats in the grid before the game starts.  Also
fix the position of the cat in the circle while game is underway.
remove numbers in the square container section of the html file as
they were being displayed on the front end when i after altering the
position of the cat.